requirements.txt
loguru==0.6.0
importlib-resources==5.4.0
Appium-Python-Client==1.1.0
